+++ oneview_redfish_toolkit/tests/test_util.py
@@ -189,8 +189,7 @@ class TestUtil(unittest.TestCase):
             self.fail('Failed to load registries files: {}'.format(e.msg))
 
     @mock.patch.object(util, 'check_oneview_availability')
-    @mock.patch.object(util, 'OneViewClient')
-    def test_load_config(self, mock_ov, check_ov_availability):
+    def test_load_config(self, check_ov_availability):
         # Test load config function
 
         util.load_config(self.config_file)
@@ -254,13 +253,10 @@ class TestUtil(unittest.TestCase):
         except Exception:
             self.fail("Failed to get a valid IP Address")
 
-    @mock.patch.object(util, 'OneViewClient')
     @mock.patch.object(util, 'check_oneview_availability')
     def test_create_certs(
-        self, check_ov_availability, oneview_client_mockup):
+        self, check_ov_availability):
         # Test generate_certificate function
-
-        check_ov_availability.return_value = 0
 
         util.load_config(self.config_file)
 
@@ -269,27 +265,20 @@ class TestUtil(unittest.TestCase):
         self.assertTrue(os.path.exists(os.path.join("certs", "test" + ".crt")))
         self.assertTrue(os.path.exists(os.path.join("certs", "test" + ".key")))
 
-    @mock.patch.object(util, 'OneViewClient')
     @mock.patch.object(util, 'check_oneview_availability')
-    @mock.patch('oneview_redfish_toolkit.util.delivery_retry_attempts', '')
-    @mock.patch('oneview_redfish_toolkit.util.delivery_retry_interval', '')
     def test_load_event_service_invalid_info(
-        self, check_ov_availability, oneview_client_mockup):
-        check_ov_availability.return_value = 0
-
+        self, check_ov_availability):
         self.assertRaises(
             OneViewRedfishError, util.load_config(self.config_file))
 
     @mock.patch.object(util, 'check_oneview_availability')
-    @mock.patch.object(util, 'OneViewClient')
     @mock.patch.object(util, 'subscriptions_by_type')
     @mock.patch(
         'oneview_redfish_toolkit.event_dispatcher.EventDispatcher.start')
     def test_submit_event_with_subscriber(
-        self, start_mock, subscription_mock, ov_mock, check_ov_availability):
+        self, start_mock, subscription_mock, check_ov_availability):
         """Tests SubmitTestEvent action with two subscribers"""
 
-        check_ov_availability.return_value = 0
         util.load_config(self.config_file)
 
         with open(
@@ -307,15 +296,13 @@ class TestUtil(unittest.TestCase):
         self.assertTrue(start_mock.call_count == 2)
 
     @mock.patch.object(util, 'check_oneview_availability')
-    @mock.patch.object(util, 'OneViewClient')
     @mock.patch.object(util, 'subscriptions_by_type')
     @mock.patch(
         'oneview_redfish_toolkit.event_dispatcher.EventDispatcher.start')
     def test_submit_event_without_subscriber(
-        self, start_mock, subscription_mock, ov_mock, check_ov_availability):
+        self, start_mock, subscription_mock, check_ov_availability):
         """Tests SubmitTestEvent action with no subscribers"""
 
-        check_ov_availability.return_value = 0
         util.load_config(self.config_file)
 
         with open(